+++ thirdParty/cupla/alpaka/include/alpaka/block/shared/st/BlockSharedMemStOmp5.hpp
@@ -15,8 +15,8 @@
 #        error If ALPAKA_ACC_ANY_BT_OMP5_ENABLED is set, the compiler has to support OpenMP 4.0 or higher!
 #    endif
 
-#    include <alpaka/block/shared/st/BlockSharedMemStMember.hpp>
 #    include <alpaka/block/shared/st/Traits.hpp>
+#    include <alpaka/block/shared/st/detail/BlockSharedMemStMemberImpl.hpp>
 
 #    include <omp.h>
 
@@ -44,10 +44,21 @@ namespace alpaka
             //-----------------------------------------------------------------------------
             static auto declareVar(BlockSharedMemStOmp5 const& smem) -> T&
             {
+                auto* data = smem.template getVarPtr<T>(TuniqueId);
+
+                if(!data)
+                {
 #    pragma omp barrier
-                smem.alloc<T>();
-#    pragma omp barrier
-                return smem.getLatestVar<T>();
+#    pragma omp single
+                    {
+                        smem.template alloc<T>(TuniqueId);
+                    }
+                    // lookup for the data chunk allocated by the master thread
+                    data = smem.template getLatestVarPtr<T>();
+                }
+                ALPAKA_ASSERT_OFFLOAD(data != nullptr);
+
+                return *data;
             }
         };
         //#############################################################################
@@ -55,9 +66,9 @@ namespace alpaka
         struct FreeSharedVars<BlockSharedMemStOmp5>
         {
             //-----------------------------------------------------------------------------
-            static auto freeVars(BlockSharedMemStOmp5 const& mem) -> void
+            static auto freeVars(BlockSharedMemStOmp5 const&) -> void
             {
-                mem.free();
+                // shared memory block data will be reused
             }
         };
     } // namespace traits