+++ thirdParty/cupla/alpaka/test/unit/core/src/OmpScheduleTest.cpp
@@ -49,7 +49,7 @@ TEST_CASE("ompSetSchedule", "[core]")
     auto const expectedSchedule = alpaka::omp::Schedule{alpaka::omp::Schedule::Dynamic, 3};
     alpaka::omp::setSchedule(expectedSchedule);
     // The check makes sense only when this feature is supported
-#if defined _OPENMP && _OPENMP >= 200805 && ALPAKA_ACC_CPU_B_OMP2_T_SEQ_ENABLED
+#if defined _OPENMP && _OPENMP >= 200805 && defined ALPAKA_ACC_CPU_B_OMP2_T_SEQ_ENABLED
     auto const actualSchedule = alpaka::omp::getSchedule();
     REQUIRE(expectedSchedule.kind == actualSchedule.kind);
     REQUIRE(expectedSchedule.chunkSize == actualSchedule.chunkSize);
@@ -64,7 +64,7 @@ TEST_CASE("ompSetNoSchedule", "[core]")
     auto const noSchedule = alpaka::omp::Schedule{alpaka::omp::Schedule::NoSchedule};
     alpaka::omp::setSchedule(noSchedule);
     // The check makes sense only when this feature is supported
-#if defined _OPENMP && _OPENMP >= 200805 && ALPAKA_ACC_CPU_B_OMP2_T_SEQ_ENABLED
+#if defined _OPENMP && _OPENMP >= 200805 && defined ALPAKA_ACC_CPU_B_OMP2_T_SEQ_ENABLED
     auto const actualSchedule = alpaka::omp::getSchedule();
     REQUIRE(expectedSchedule.kind == actualSchedule.kind);
     REQUIRE(expectedSchedule.chunkSize == actualSchedule.chunkSize);