+++ thirdParty/cupla/alpaka/test/unit/kernel/src/KernelWithOmpSchedule.cpp
@@ -1,4 +1,4 @@
-/* Copyright 2020 Sergei Bastrakov
+/* Copyright 2020-2021 Sergei Bastrakov
  *
  * This file is part of alpaka.
  *
@@ -7,6 +7,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+#include <alpaka/core/BoostPredef.hpp>
 #include <alpaka/core/OmpSchedule.hpp>
 #include <alpaka/kernel/Traits.hpp>
 #include <alpaka/test/KernelExecutionFixture.hpp>
@@ -17,9 +18,6 @@
 #include <cstdint>
 
 //#############################################################################
-// Schedule to be used by all kernels in this file
-static constexpr auto expectedSchedule = alpaka::omp::Schedule{alpaka::omp::Schedule::Dynamic, 10};
-
 // Base kernel, not to be used directly in unit tests
 struct KernelWithOmpScheduleBase
 {
@@ -28,68 +26,70 @@ struct KernelWithOmpScheduleBase
     template<typename TAcc>
     ALPAKA_FN_ACC auto operator()(TAcc const& acc, bool* success) const -> void
     {
-        // By default no run-time check is performed
+        // No run-time check is performed
         alpaka::ignore_unused(acc);
         ALPAKA_CHECK(*success, true);
     }
-
-    // Only check when the schedule feature is active
-#if defined _OPENMP && _OPENMP >= 200805 && ALPAKA_ACC_CPU_B_OMP2_T_SEQ_ENABLED
-    ALPAKA_NO_HOST_ACC_WARNING
-    template<typename TDim, typename TIdx>
-    ALPAKA_FN_ACC auto operator()(alpaka::AccCpuOmp2Blocks<TDim, TIdx> const& acc, bool* success) const -> void
-    {
-        alpaka::ignore_unused(acc);
-        omp_sched_t kind;
-        int actualChunkSize = 0;
-        omp_get_schedule(&kind, &actualChunkSize);
-        auto const actualKind = static_cast<std::uint32_t>(kind);
-        bool result = (expectedSchedule.kind == actualKind) && (expectedSchedule.chunkSize == actualChunkSize);
-        ALPAKA_CHECK(*success, result);
-    }
-#endif
 };
 
-// Kernel that sets the schedule via constexpr ompSchedule.
-// Checks that this variable is only declared and not defined, It also tests that
-// alpaka never odr-uses it.
-struct KernelWithConstexprMemberOmpSchedule : KernelWithOmpScheduleBase
+// Kernel that sets the schedule kind via constexpr ompScheduleKind, but no chunk size.
+// Checks that this variable is only declared and not defined, It also tests that alpaka never odr-uses it.
+template<alpaka::omp::Schedule::Kind TKind>
+struct KernelWithConstexprMemberOmpScheduleKind : KernelWithOmpScheduleBase
 {
-    static constexpr auto ompSchedule = expectedSchedule;
+    static constexpr auto ompScheduleKind = TKind;
 };
 
-// Kernel that sets the schedule via non-constexpr ompSchedule.
-struct KernelWithMemberOmpSchedule : KernelWithOmpScheduleBase
+// Kernel that sets the schedule kind via non-constexpr ompScheduleKind, but no chunk size.
+template<alpaka::omp::Schedule::Kind TKind>
+struct KernelWithMemberOmpScheduleKind : KernelWithOmpScheduleBase
 {
-    static const alpaka::omp::Schedule ompSchedule;
+    static const alpaka::omp::Schedule::Kind ompScheduleKind = TKind;
+};
+
+// Kernel that sets the schedule chunk size via constexpr ompScheduleChunkSize in addition to schedule kind, but no
+// chunk size. Checks that this variable is only declared and not defined, It also tests that alpaka never odr-uses it.
+template<alpaka::omp::Schedule::Kind TKind>
+struct KernelWithConstexprStaticMemberOmpScheduleChunkSize : KernelWithConstexprMemberOmpScheduleKind<TKind>
+{
+    static constexpr int ompScheduleChunkSize = 5;
+};
+
+// Kernel that sets the schedule chunk size via non-constexpr ompScheduleChunkSize in addition to schedule kind.
+template<alpaka::omp::Schedule::Kind TKind>
+struct KernelWithStaticMemberOmpScheduleChunkSize : KernelWithMemberOmpScheduleKind<TKind>
+{
+    static const int ompScheduleChunkSize;
 };
 // In this case, the member has to be defined externally
-const alpaka::omp::Schedule KernelWithMemberOmpSchedule::ompSchedule = expectedSchedule;
+template<alpaka::omp::Schedule::Kind TKind>
+const int KernelWithStaticMemberOmpScheduleChunkSize<TKind>::ompScheduleChunkSize = 2;
 
-// Kernel that sets the schedule via partial specialization of a trait
-struct KernelWithTraitOmpSchedule : KernelWithOmpScheduleBase
+// Kernel that sets the schedule chunk size via non-constexpr non-static ompScheduleChunkSize in addition to schedule
+// kind.
+template<alpaka::omp::Schedule::Kind TKind>
+struct KernelWithMemberOmpScheduleChunkSize : KernelWithConstexprMemberOmpScheduleKind<TKind>
 {
+    int ompScheduleChunkSize = 4;
 };
 
-// Kernel that sets the schedule via both member and partial specialization of a trait.
-// In this case test that the trait is used, not the member.
-struct KernelWithMemberAndTraitOmpSchedule : KernelWithOmpScheduleBase
+// Kernel that copies the given base kernel and adds an OmpSchedule trait on top
+template<typename TBase>
+struct KernelWithTrait : TBase
 {
-    // Set to be different from expected so that it this is used the test would fail
-    static constexpr auto ompSchedule = alpaka::omp::Schedule{expectedSchedule.kind, expectedSchedule.chunkSize + 1};
 };
 
 namespace alpaka
 {
     namespace traits
     {
-        // Specialize the trait for all kernels
-        template<typename TKernelFnObj, typename TAcc>
-        struct OmpSchedule<TKernelFnObj, TAcc>
+        // Specialize the trait for kernels of type KernelWithTrait<>
+        template<typename TBase, typename TAcc>
+        struct OmpSchedule<KernelWithTrait<TBase>, TAcc>
         {
             template<typename TDim, typename... TArgs>
             ALPAKA_FN_HOST static auto getOmpSchedule(
-                TKernelFnObj const& kernelFnObj,
+                KernelWithTrait<TBase> const& kernelFnObj,
                 Vec<TDim, Idx<TAcc>> const& blockThreadExtent,
                 Vec<TDim, Idx<TAcc>> const& threadElemExtent,
                 TArgs const&... args) -> alpaka::omp::Schedule
@@ -99,7 +99,7 @@ namespace alpaka
                 alpaka::ignore_unused(threadElemExtent);
                 alpaka::ignore_unused(args...);
 
-                return expectedSchedule;
+                return alpaka::omp::Schedule{alpaka::omp::Schedule::Static, 4};
             }
         };
     } // namespace traits
@@ -107,7 +107,7 @@ namespace alpaka
 
 // Generic testing routine for the given kernel type
 template<typename TAcc, typename TKernel>
-void test()
+void testKernel()
 {
     using Acc = TAcc;
     using Dim = alpaka::Dim<Acc>;
@@ -115,31 +115,47 @@ void test()
 
     alpaka::test::KernelExecutionFixture<Acc> fixture(alpaka::Vec<Dim, Idx>::ones());
 
+    // Base version with no OmpSchedule trait
     TKernel kernel;
-
     REQUIRE(fixture(kernel));
+
+    // Same members, but with OmpSchedule trait
+    KernelWithTrait<TKernel> kernelWithTrait;
+    REQUIRE(fixture(kernelWithTrait));
 }
 
-//-----------------------------------------------------------------------------
-TEMPLATE_LIST_TEST_CASE("kernelWithConstexprMemberOmpSchedule", "[kernel]", alpaka::test::TestAccs)
+// Note: it turned out not possible to test all possible combinations as it causes several compilers to crash in CI.
+// However the following tests should cover all important cases
+
+TEMPLATE_LIST_TEST_CASE("kernelWithOmpScheduleBase", "[kernel]", alpaka::test::TestAccs)
 {
-    test<TestType, KernelWithConstexprMemberOmpSchedule>();
+    testKernel<TestType, KernelWithOmpScheduleBase>();
 }
 
-//-----------------------------------------------------------------------------
-TEMPLATE_LIST_TEST_CASE("kernelWithMemberOmpSchedule", "[kernel]", alpaka::test::TestAccs)
+TEMPLATE_LIST_TEST_CASE("kernelWithConstexprMemberOmpScheduleKind", "[kernel]", alpaka::test::TestAccs)
 {
-    test<TestType, KernelWithMemberOmpSchedule>();
+    testKernel<TestType, KernelWithConstexprMemberOmpScheduleKind<alpaka::omp::Schedule::NoSchedule>>();
 }
 
-//-----------------------------------------------------------------------------
-TEMPLATE_LIST_TEST_CASE("kernelWithTraitOmpSchedule", "[kernel]", alpaka::test::TestAccs)
+TEMPLATE_LIST_TEST_CASE("kernelWithMemberOmpScheduleKind", "[kernel]", alpaka::test::TestAccs)
 {
-    test<TestType, KernelWithTraitOmpSchedule>();
+    testKernel<TestType, KernelWithMemberOmpScheduleKind<alpaka::omp::Schedule::Static>>();
 }
 
-//-----------------------------------------------------------------------------
-TEMPLATE_LIST_TEST_CASE("kernelWithMemberAndTraitOmpSchedule", "[kernel]", alpaka::test::TestAccs)
+TEMPLATE_LIST_TEST_CASE("kernelWithConstexprStaticMemberOmpScheduleChunkSize", "[kernel]", alpaka::test::TestAccs)
 {
-    test<TestType, KernelWithMemberAndTraitOmpSchedule>();
+    testKernel<TestType, KernelWithConstexprStaticMemberOmpScheduleChunkSize<alpaka::omp::Schedule::Dynamic>>();
+}
+
+TEMPLATE_LIST_TEST_CASE("kernelWithStaticMemberOmpScheduleChunkSize", "[kernel]", alpaka::test::TestAccs)
+{
+    testKernel<TestType, KernelWithStaticMemberOmpScheduleChunkSize<alpaka::omp::Schedule::Guided>>();
+}
+
+TEMPLATE_LIST_TEST_CASE("kernelWithMemberOmpScheduleChunkSize", "[kernel]", alpaka::test::TestAccs)
+{
+#if defined _OPENMP && _OPENMP >= 200805
+    testKernel<TestType, KernelWithMemberOmpScheduleChunkSize<alpaka::omp::Schedule::Auto>>();
+#endif
+    testKernel<TestType, KernelWithMemberOmpScheduleChunkSize<alpaka::omp::Schedule::Runtime>>();
 }