+++ thirdParty/cupla/include/cupla/namespace.hpp
@@ -54,8 +54,8 @@
 #           define CUPLA_ACCELERATOR_NAMESPACE cupla_tbb_seq_async
 #       endif
 
-#       ifdef ALPAKA_ACC_CPU_BT_OMP4_ENABLED
-#           define CUPLA_ACCELERATOR_NAMESPACE cupla_omp4_omp4_async
+#       ifdef ALPAKA_ACC_ANY_BT_OMP5_ENABLED
+#           define CUPLA_ACCELERATOR_NAMESPACE cupla_omp5_omp5_async
 #       endif
 
 #   endif // mixed accelerator usage
@@ -91,8 +91,8 @@
 #           define CUPLA_ACCELERATOR_NAMESPACE cupla_tbb_seq_sync
 #       endif
 
-#       ifdef ALPAKA_ACC_CPU_BT_OMP4_ENABLED
-#           define CUPLA_ACCELERATOR_NAMESPACE cupla_omp4_omp4_sync
+#       ifdef ALPAKA_ACC_ANY_BT_OMP5_ENABLED
+#           define CUPLA_ACCELERATOR_NAMESPACE cupla_omp5_omp5_sync
 #       endif
 
 #   endif // mixed accelerator usage