+++ thirdParty/cupla/alpaka/CHANGELOG.md
@@ -3,6 +3,33 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.6.1] - 2021-06-29
+### Compatibility Changes:
+- rework implementation of OpenMP schedule support #1279 #1309 #1313 #1341
+  - `alpaka::omp::Schedule` is replaced by `ompScheduleKind` and `ompScheduleChunkSize`
+### Bug Fixes:
+- fix OpenMP 5 shared memory allocation #1254 
+- fix static shared memory alignment #1282
+- fix BlockSharedMemStMemberImpl::getVarPtr for last var #1280
+- fix CPU static shared memory implementation #1258
+- unit tests: fix queue test #1266
+- fix CtxBlockOacc: SyncBlockThreads #1291
+- fix assert in DeclareSharedVar (OpenAcc) #1303
+- CMake CUDA: dev compile options not propagated #1294
+- example: fix warning (NVCC+OpenMP) #1307
+- TBB: Add missing <limits> header and fix integer namespace #1327
+- OpenAcc: TaskKernelOacc: copyin(all used local vars) #1342
+- port macOSX CI fix from #1283
+- CI: use ubuntu-18.04 for gcc-5 and gcc-6 builds #1252
+- CI: disable GCC 10.3 + NVCC tests #1302
+- CI: MSVC + nvcc workarounds and fixes #1332
+- CI: fix warp test #1339
+
+### Misc
+- add ALPAKA_ASSERT_OFFLOAD Macro #1260
+- document return value of `empty()` and `isComplete()` #1265
+- Prefer TBBConfig.cmake over FindTBB.cmake #1329
+
 
 ## [0.6.0] - 2021-01-20
 ### Compatibility Changes: